ssh-agent fixes with max buffer checks and unit tests

+++ lib/Pogo/Util/SSH/Agent.pm
@@ -6,11 +6,12 @@ use warnings;
 use Log::Log4perl qw(:easy);
 use AnyEvent;
 use AnyEvent::Strict;
-use Pogo::Object::Event;
 use AnyEvent;
 use AnyEvent::Util qw( run_cmd );
 use File::Temp qw( tempfile );
-use POSIX qw( mkfifo O_NONBLOCK O_RDONLY O_WRONLY );
+use POSIX qw( mkfifo O_NONBLOCK O_RDONLY O_WRONLY PIPE_BUF );
+
+use base 'Object::Event';
 
 ###########################################
 sub new {
@@ -18,27 +19,31 @@ sub new {
     my($class, %options) = @_;
 
     my $self = {
-        fifo_path       => "/tmp/fifo",
-        fifo_perms      => 0600,
-        ssh_agent       => "ssh-agent",
-        ssh_add         => "ssh-add",
-        tempdir         => undef,
-        ssh_add_timeout => 10,
+        fifo_perms        => 0600,
+        tempdir           => undef,
+        ssh_agent         => "ssh-agent",
+        ssh_add           => "ssh-add",
+        ssh_add_timeout   => 60,
+        ssh_agent_timeout => 60,
+        started           => 0,
+        fifo_opened       => 0,
         %options,
     };
 
     bless $self, $class;
+
+    return $self;
 }
 
 ###########################################
 sub start {
 ###########################################
-    my( $self, $cb ) = @_;
+    my( $self ) = @_;
 
     my @dir = ();
     @dir = ( DIR => $self->{ tempdir } ) if defined $self->{ tempdir };
 
-    my ($fh, $filename) = tempfile( @dir );
+    my ($fh, $filename) = tempfile( @dir, UNLINK => 1 );
     unlink $filename or LOGDIE "$!";
     $self->{ socket } = $filename;
 
@@ -47,6 +52,13 @@ sub start {
     # $ ssh-agent -d
     # SSH_AUTH_SOCK=/tmp/ssh-pFEHe26945/agent.26945; export SSH_AUTH_SOCK;
     # echo Agent pid 26945;
+    #
+    # Problem is that ssh-agent checks if it's running in a terminal and
+    # if not, will suppress all output. We can either use a pty and grab
+    # its output to find the unix socket the agent is listenting on, or
+    # force a pre-determined file. We're doing the latter for now until
+    # I've figured out how to convince run_cmd to use a pty conduit like
+    # POE::Wheel::Run does.
 
     my $cmd = run_cmd [ $self->{ ssh_agent }, "-d", "-a", $filename ],
         "<",  "/dev/null",
@@ -55,37 +67,32 @@ sub start {
         '$$', \$self->{ pid },
         ;
 
-    $cmd->cb( sub {
-        my( $c ) = @_;
-
-        $c->recv() and 
-            LOGDIE "$self->{ ssh_agent } failed ($!)";
-    } );
-
-        # Wait and poll until socket exists
+        # Wait and poll until ssh-agent unix socket exists
     my $cv = AnyEvent->condvar();
-    my $ssh_add_starttime = time();
+    my $ssh_agent_starttime = time();
 
     my $timer; $timer = AnyEvent->timer( 
         after    => 0, 
-        interval => 1,
+        interval => 0.5,
         cb       => sub {
+
+            DEBUG "checking for to-be-created ssh-agent socket $filename";
             if( -e $filename ) {
-                $cv->send( 1 );
-            }
-            if( time() - $ssh_add_starttime >
-                $self->{ ssh_add_timeout } ) {
-                DEBUG "Tired of waiting for ssh-add";
-                $cv->send( 0 );
                 undef $timer;
+                DEBUG "ssh-agent created unix socket: $filename";
+                $self->event( "ssh_agent_ready", $filename, $self->{ pid } );
+                $self->{ started } = 1;
+            }
+
+            if( time() - $ssh_agent_starttime >
+                $self->{ ssh_agent_timeout } ) {
+                DEBUG "Tired of waiting for $self->{ ssh_agent }";
+                $self->fifo_cleanup();
+                undef $timer;
+                $self->event( "ssh_agent_error", "timeout" );
             }
         } 
     );
-
-    if( $cv->recv() ) {
-        DEBUG "Socket $filename exists";
-        $cb->( $filename, $self->{ pid } );
-    }
 }
 
 ###########################################
@@ -93,13 +100,35 @@ sub key_add {
 ###########################################
     my( $self, $key ) = @_;
 
-    DEBUG "Adding key to fifo";
+    my $keylen = length( $key );
+
+    DEBUG "Adding key (len=$keylen) to fifo";
+
+    if( $keylen > PIPE_BUF ) {
+        ERROR "Key too long (max is ", PIPE_BUF, ")";
+        $self->event( "ssh_key_added_fail" );
+    }
+
     $self->fifo_setup( $key );
 
     $ENV{ SSH_AUTH_SOCK } = $self->{ socket };
-    my $cmd = run_cmd [ $self->{ ssh_add }, $self->{ fifo_path } ];
-    $cmd->recv();
-    DEBUG "Key added to fifo";
+    my $cmd = run_cmd [ $self->{ ssh_add }, $self->{ fifo_path } ],
+    "2>" => "/dev/null",
+       # useless because ssh-add suppresses messages unless it's 
+       # on a tty, but at least suppresses tty output during testing
+    ;
+
+    $cmd->cb( sub {
+        my $rc = $cmd->recv();
+
+        if( $rc == 0 ) {
+            DEBUG "Key added to fifo";
+            $self->event( "ssh_key_added_ok" );
+        } else {
+            ERROR "$self->{ ssh_add } reported error: $!";
+            $self->event( "ssh_key_added_fail" );
+        }
+    } );
 }
 
 ###########################################
@@ -108,6 +137,14 @@ sub socket {
     my( $self ) = @_;
 
     return $self->{ socket };
+}
+
+###########################################
+sub fifo_path {
+###########################################
+    my( $self ) = @_;
+
+    return $self->{ fifo_path };
 }
 
 ###########################################
@@ -123,7 +160,7 @@ sub fifo_setup {
     my @dir = ();
     @dir = ( DIR => $self->{ tempdir } ) if defined $self->{ tempdir };
 
-    my ($fh, $filename) = tempfile( @dir );
+    my ($fh, $filename) = tempfile( @dir, UNLINK => 1 );
     unlink $filename or LOGDIE "$!";
 
     mkfifo $filename, $self->{ fifo_perms } or 
@@ -145,6 +182,8 @@ sub fifo_setup {
     $self->{ fifo_r_fd } = $fdr;
     $self->{ fifo_path } = $filename;
 
+    $self->{ fifo_opened } = 1;
+
     syswrite $fdw, $key;
 
     DEBUG "Bytes left: ", fifo_bytecount( $fdr ), "\n";
@@ -154,7 +193,7 @@ sub fifo_setup {
         interval => 1,
         cb => sub {
             my $bytes_left = fifo_bytecount( $fdr );
-            DEBUG "Timer: Bytes left in pipe: $bytes_left";
+            # DEBUG "Timer: Bytes left in pipe: $bytes_left";
     
             if( $bytes_left == 0 ) {
                 DEBUG "Refilling buffer";
@@ -172,13 +211,20 @@ sub fifo_cleanup {
 ###########################################
     my( $self ) = @_;
 
+    if( ! $self->{ fifo_opened } ) {
+        DEBUG "Fifo was never opened";
+        return 1;
+    }
+
       # stop refreshing the pipe
     $self->{ fifo_refresher } = undef;
 
-    close $self->{ fifo_w_fh };
-    close $self->{ fifo_r_fh };
+    close $self->{ fifo_w_fd };
+    close $self->{ fifo_r_fd };
 
     unlink $self->{ fifo_path };
+
+    $self->{ fifo_opened } = 0;
 }
 
 ###########################################
@@ -197,21 +243,125 @@ sub shutdown {
 ###########################################
     my( $self ) = @_;
 
-    kill 2, $self->{ pid } or 
-        die "Can't kill pid $self->{ pid } ($!)";
-
-    if( -e $self->{ socket } ) {
-        unlink $self->{ socket } or
-            die "Can't unlink $self->{ socket } ($!)";
+    if( ! $self->{ started } ) {
+        DEBUG "Already shut down";
+        return 1;
     }
+
+    DEBUG "Shutting down";
+
+    $self->fifo_cleanup();
+
+    if( kill 0, $self->{ pid } ) {
+        DEBUG "$$: $self->{ pid } is running, sending signal";
+        if( !kill 2, $self->{ pid } ) {
+            ERROR "Can't kill pid $self->{ pid } ($!)";
+            return undef;
+        }
+        INFO "Waiting for $self->{ ssh_agent } (pid $self->{ pid }) to ",
+          "terminate ...";
+        waitpid( $self->{ pid }, 0 );
+        INFO "$self->{ pid } is gone.";
+    }
+
+    for my $file ( $self->{ socket }, $self->{ fifo_path } ) {
+        next if !defined $file;
+        next if !-e $file;
+
+        DEBUG "Removing $file.";
+        if( !unlink $file ) {
+            ERROR "Can't unlink $file ($!)";
+            return undef;
+        }
+    }
+
+    DEBUG "Shutdown complete.";
+    $self->event( "shutdown_complete" );
+
+    $self->{ started } = 0;
+    return 1;
 }
 
 ###########################################
-sub DESTROY {
+sub test_run {
 ###########################################
-    my( $self ) = @_;
+    my( $bin_dir ) = @_;
 
-    $self->shutdown();
+    require Test::More;
+    require Sysadm::Install;
+
+    my $agent = Pogo::Util::SSH::Agent->new();
+
+    my $cv = AnyEvent->condvar();
+
+    my $g1 = $agent->reg_cb( "ssh_key_added_ok", sub {
+        my( $message ) = @_;
+
+        Test::More::ok( 1, "key added" );
+
+        DEBUG "*** fifo path is ", $agent->fifo_path();
+        DEBUG "*** socket is ", $agent->socket();
+        DEBUG "SSH_AUTH_SOCK=", $agent->socket(), " ssh localhost";
+
+        $cv->send();
+    } );
+
+    my $g2 = $agent->reg_cb( "ssh_key_add_failed", sub {
+            Test::More::ok( 0, "key added" );
+    } );
+
+    my $g3 = $agent->reg_cb( "shutdown_complete", sub {
+            Test::More::ok( 1, "shutdown_complete" );
+    } );
+
+    my $g4 = $agent->reg_cb( "ssh_agent_ready", sub {
+        my( $auth_sock, $agent_pid ) = @_;
+
+        Test::More::ok( 1, "auth socket reported" );
+
+        DEBUG "auth_sock is $auth_sock";
+        $agent->key_add( test_privkey() );
+    } );
+
+    $agent->start();
+
+    $cv->recv();
+
+    $agent->shutdown();
+}
+
+###########################################
+sub test_privkey {
+###########################################
+    return <<EOT;
+-----BEGIN RSA PRIVATE KEY-----
+MIIEogIBAAKCAQEA1fbGvcRmGy6oVrtbVqqjMM3LEVgUkzvCZhsLxdj2ptvWex8U
+vatlYUqtdlKg03dwf8XBdi4zLOpu6zPEFi8XvaxaBeBz7NxmF9CDIzuWw1NM+mZP
+ukGPYyzXWDWfVGLJGFYtqrAYtKoAaO+46EW2V2sfaTWcCjUdvGVl0/7klfiNnR80
+iwBbuv0UOCVE4D7Vz3phUHuVGFs/1FLCOyXVZyDrQPHVkW4AT1RjDjMyHm0w/cs1
+YI2wKU0kKKFdHmBbQXw2twMWZbCteYUENRASh/BUeNVNDznqlrFSID+38KiMqT/O
+u1qImTOehXsOiXlI9i2ZCfCmEmEAYq0LQC838QIBIwKCAQBn7OQwSXNsSdy8aaFk
+nAYfBN75y7I44oL+ZOh2CkvqpUrrWD1GLq2Vp+3aYqXjDiCzFujwQlNe9YZU++Lm
+NCF5YlecdFWQTcswI3LlOjNI7fIwetZEhj5UvgIyKKx5cc9jl5KGGwSvhcWvT914
+Idw5FsYdKKrgYvEvnvbx8NVtaTU+mWxoRe0PYXMsiI3oEvSYl9HLkQUqbhN1Y5T+
+Zuoog9i0c+cTfgDFqoCB5JA5rzAB/5o3cwlg9urPMIDbf0tE1hsaFuNwuKR8x7QY
+RLW1shULyye9wO6VAxnQwS68mXhjJEVJFCa3himwWf7RRC6zQ5dlf31MJQis8REG
+g/vbAoGBAPZiAfRfHS/wbwj0qu1NuAPRlGyMINaK2JdV1GABBkLmDg5wsLrGgK6C
+EVuiP2yQ9vPReR5jlkoZ3lC+PH0yJ8hvvQEC0HV+WI++jyMTYvpTptYcEGs/U7Pm
+7RGhDCsZiPbmkGFum46b5kQkiDj7HrNENSYHi+CpDTWrdgSMhlYFAoGBAN5Q0lHr
+y60jXAAF6Lrw9mHw6Z8wTkqBctrJC7cBLZJLxy23Fj8MNTIYL0okSmBh/n/7B50/
+gwd0+f8LHyHsFGFEBmdRTsyk6wdQpgTvtjivJiXbE4/975bIRWWzS3fpayVvCmYu
+wqpHOSKqk+cMsBduEvAEQCzbniJEGYlL5HH9AoGAd6vyUh+R1XTIN4zIDNybNQ4G
+Q1oBUkNwhAUeAr6rRRCnvd72wR6WRiHrLIIBjIDs+hVJdSkOe8Nr+1UWEOx5uSBU
+fNV7McEGcbRUJvrJrMmLjJFJzbELZgJzJdHhVsNCho0+0D0Juku4/IbF0oiZ4gsv
+wgOqVy2KEsD+zwJtIncCgYEA1/a9rqqLV7vy+LVIexX2qEkddhGrI841Dwxxx7gA
+Yjr8AIX4WoDjN/o8kSqRZPF64rlX2pV3+J2FI6RnYsgTzDNz71ZMjEhvSO9CMK5Z
+PmEAfIusmoGm6j7kVCqD05JK0+g1/Nz3nhlNciIMBQUC1O6WDbr8g1kABAejxzPH
++bMCgYEAhtn+/SlcPRSbh90Ghv2ys67ga3cTFj+hXCotVkvfGGu5nA29QXxkka2y
+xx//lsBuHAHFq2rcqktz4VERCgg4nT3v33j13vRFEZpkMl33aYXnc1vvIur8NZHE
+hNV0NLsHq/b8VjnRzNBWqjKQXeIUurGtt2Bso/xGkYRFEfUim88=
+-----END RSA PRIVATE KEY-----
+EOT
 }
 
 1;
@@ -228,17 +378,20 @@ Pogo::Util::SSH::Agent - Start ssh-agent and add keys to it
 
     my $agent = Pogo::Util::SSH::Agent->new();
 
-    $agent->start( sub {
+    my $cv = AnyEvent->condvar();
+
+    $agent->reg_cb( "ssh_agent_ready", sub {
         my( $auth_sock, $agent_pid ) = @_;
 
+        $agent->key_add( $private_key );
         # ...
     } );
 
-    $agent->key_add( $private_key, sub {
-        my( $rc ) = @_;
+    $agent->reg_cb( "ssh_key_added_ok", sub {
+        $cv->send();
+    };
 
-        # ...
-    } );
+    $cv->recv();
 
     $agent->shutdown();
 
@@ -250,17 +403,24 @@ socket the agent is listening on, which can be used later when invoking
 the C<ssh> command, which will look in an environment variable name
 C<$SSH_AUTH_SOCK> for the unix socket file path.
 
-=head1 METHODS
+=head2 METHODS
 
 =over 4
 
 =item C<new()>
 
-Constructor.
+Constructor, uses the following parameters as defaults, override if
+you want something else:
+
+        fifo_perms      => 0600,          # permissions settings for fifo
+        ssh_agent       => "ssh-agent",   # path to ssh-agent
+        ssh_add         => "ssh-add",     # path to ssh-add
+        tempdir         => undef,         # fifo dir
+        ssh_add_timeout => 10,            # max wait time for ssh-add 
 
 =item C<auth_socket()>
 
-Set/get the auth socket the agent is listening on.
+Set/get the auth socket file the agent is listening on.
 
 =item C<start()>
 
@@ -274,6 +434,34 @@ Employs some crazy fifo logic to accomplish this.
 =item C<shutdown()>
 
 Shuts down the ssh-agent process.
+
+=back
+
+=head2 EVENTS
+
+=over 4
+
+=item C<ssh_agent_ready>
+
+C<ssh-agent> was started and has created the unix socket. Event
+carries the ssh-agent's pid as parameter.
+
+=item C<ssh_agent_error>
+
+If the ssh-agent returns an error, this event is emitted with the Unix
+error ($!) as an argument.
+
+=item C<ssh_agent_done>
+
+The ssh-agent has been shut down successfully.
+
+=item C<ssh_key_added_ok>
+
+The key has been added to the agent.
+
+=item C<ssh_key_added_fail>
+
+The agent return an error when we tried to add the key.
 
 =back
